attention: Remove unnecessary contiguous operations
Since we are permuting both query and key in the same way, we can
matmul them without first making them contiguous. Avoiding these
operations increases performance by 10% on both Metal and CUDA.

This moves the permutations into the attention operation, which
provides a better interface to callers as the inputs are now consistent
with the outputs. Of course, it also makes it easier to introduce
optimizations such as this one or to support kernels with different
requirements.